Некоторые изменения с настройками файловой директории

+++ include/_func.php
@@ -166,7 +166,7 @@ function stable($table){
 			if (strlen($table['_where'])) $sql .= " AND {$table['_where']}";
 			if ($table['debug']) mpre($sql);
 			$result = mpql(mpqw($sql), 0);
-			if(array_key_exists('img', $result) && ($file_name = mpopendir("include/{$result['img']}"))){
+			if(!empty($result['img']) && /*array_key_exists('img', $result) &&*/ ($file_name = mpopendir("include/{$result['img']}"))){
 				@mkdir("/tmp/del/", true);
 				copy($file_name, "/tmp/del/". basename($file_name));
 				unlink($file_name);
@@ -190,7 +190,7 @@ mpre($_GET); exit;
 		}elseif(count($_GET['inc']) || count($_GET['dec'])){
 			$m = ($_GET['dec'] > $_GET['inc'] ? 'dec' : 'inc');
 			list($k) = each($_GET[$m]);
-			$sql = "SELECT id, $k FROM {$table['name']} WHERE $k ".($m == 'inc' ? '>=' : '<=')." (SELECT $k FROM {$table['name']} WHERE id = {$_GET[$m][$k]})";
+			$sql = "SELECT id, $k FROM {$table['name']} WHERE $k ".($m == 'inc' ? '>=' : '<=')." (SELECT `$k` FROM {$table['name']} WHERE id=". (int)$_GET[$m][$k]. ")";
 			if (strlen($table['where'])) $sql .= " AND ".mpquot($table['where']);
 			if (strlen($table['_where'])) $sql .= " AND {$table['_where']}";
 			$sql .= " ORDER BY $k".($m == 'dec' ? '  DESC' : '')." LIMIT 2";
@@ -207,9 +207,10 @@ mpre($_GET); exit;
 		if((substr($_fields, -3, 3) == "_id") && ($f = substr($_fields, 0, -3)) && !empty($conf['settings']["{$arg['modpath']}_{$f}"])){
 			$table['etitle'][ $_fields ] = $conf['settings']["{$arg['modpath']}_{$f}"];
 			if(substr($conf['settings'][ substr($_GET['r'], strlen($conf['db']['prefix']), 999) ], 0, 1) == '.'){
-				if(empty($table['spisok'][ $_fields ]))
+				if(empty($table['spisok'][ $_fields ])){
 					$table['spisok'][ $_fields ] = array('*'=>array('')+spisok("SELECT id, name FROM {$conf['db']['prefix']}{$arg['modpath']}_{$f} ORDER BY name"));
-				$table['shablon'][ $_fields ] = array("*"=>"<a href=\"/?m[{$arg['modpath']}]=admin&r={$conf['db']['prefix']}{$arg['modpath']}_{$f}&where[id]={f:{f}}\">{spisok:{f}}</a>");
+					$table['shablon'][ $_fields ] = array("*"=>"<a href=\"/?m[{$arg['modpath']}]=admin&r={$conf['db']['prefix']}{$arg['modpath']}_{$f}&where[id]={f:{f}}\">{spisok:{f}}</a>");
+				}
 			}
 		}
 	}
